% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_stats.R
\name{compute_stats}
\alias{compute_stats}
\title{Compute summary statistics for a raster layer}
\usage{
compute_stats(bio1_masked)
}
\arguments{
\item{bio1_masked}{A masked `SpatRaster`.}
}
\value{
A `data.frame` with min, max, mean, and sd columns.
}
\description{
Compute summary statistics for a raster layer
}
