% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_south_america.R
\name{crop_south_america}
\alias{crop_south_america}
\title{Crop a raster to the South American bounding box}
\usage{
crop_south_america(bio1)
}
\arguments{
\item{bio1}{A `SpatRaster` layer.}
}
\value{
A `SpatRaster` cropped to the extent of South America.
}
\description{
Crop a raster to the South American bounding box
}
