% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bioclim.R
\name{download_bioclim}
\alias{download_bioclim}
\title{Download WorldClim bioclimatic data}
\usage{
download_bioclim(cache_path)
}
\arguments{
\item{cache_path}{Character. Directory path for caching downloaded data.}
}
\value{
A `SpatRaster` stack of 19 bioclimatic variables.
}
\description{
Download WorldClim bioclimatic data
}
