% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_boundaries.R
\name{load_boundaries}
\alias{load_boundaries}
\title{Load South American country boundaries}
\usage{
load_boundaries(cache_path)
}
\arguments{
\item{cache_path}{Character. Directory used for caching.}
}
\value{
A `SpatVector` of the merged South American continent polygon.
}
\description{
Downloads and merges GADM level-0 polygons for all South American countries.
Results are cached to `cache_path` as an RDS file.
}
