% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_bioclim.R
\name{mask_bioclim}
\alias{mask_bioclim}
\title{Mask a raster to South American boundaries}
\usage{
mask_bioclim(bio1_cropped, sa_polygons)
}
\arguments{
\item{bio1_cropped}{A `SpatRaster` cropped to South America's extent.}

\item{sa_polygons}{A `SpatVector` of South American boundaries.}
}
\value{
A masked `SpatRaster`.
}
\description{
Reprojects `sa_polygons` to match `bio1_cropped` if CRS differs, then masks.
}
