% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the nicheRshiny Shiny application}
\usage{
run_app()
}
\value{
Launches a Shiny app; does not return a value.
}
\description{
Launch the nicheRshiny Shiny application
}
\examples{
if (interactive()) run_app()
}
